% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BMP.R
\docType{data}
\name{BMP}
\alias{BMP}
\title{Bandiagara Domain Boundary Data}
\format{\code{BMP} is a list of three dataframes including vertices information of external and internal boundaries of Bandiagara.
\describe{
  \item{bound}{The coordinates of external boundaries.}
  \item{H1}{The coordinates of an internal hole in the region.}
  \item{H2}{The coordinates of an internal hole in the region.}
}}
\usage{
data('BMP')
}
\description{
A list of coordinates of the boundary vertices  of Bandiagara
}
\references{
Data courtesy by Jean Gaudart, 2012.
}
\keyword{datasets}
