\name{horseshoe}
\alias{horseshoe}
\docType{data}
%- Also NEED an `\alias' for EACH other topic documented here.
\title{Horseshore Domain Boundary Data}
\description{This dataset includes the boundary data from horseshore domain.}
\usage{
data("horseshoe")}
%- maybe also `usage' for other objects documented here.

\format{ \code{horseshoe} is a matrix with 30 rows and two columns:
\describe{
\item{V1}{The longtitude of a boudary point.}
\item{V2}{The lattitude of a boudary point}
}}




\examples{
## see ?TriMesh help files
}
