\name{mymontreal}
\alias{mymontreal}
\docType{data}
%- Also NEED an `\alias' for EACH other topic documented here.
\title{Montreal Domain Boundary Data}
\description{This dataset includes the boundary data from Montreal domain.}
\usage{
data("mymontreal")}
%- maybe also `usage' for other objects documented here.

\format{ \code{mymontreal} is a list f three dataframes including vertices information of external and internal boundaries of Montreal.
\describe{
\item{bound}{The coordinates of external boundaries.}
\item{H1}{The coordinates of an internal hole in the region.}
\item{H2}{The coordinates of an internal hole in the region.}
}}

\examples{
## see ?TriMesh help files
}
