% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TriMesh.R
\name{TriMesh}
\alias{TriMesh}
\title{Create Triangles Mesh in 2D Domains}
\usage{
TriMesh(Pt, n, H = NULL)
}
\arguments{
\item{Pt}{A two by \code{N} matrix which indicates the outer boundry points of a 2D region.}

\item{n}{An integer parameter controlling the fineness of the triangulation
and subsequent triangulation. As n increases the fineness increases. Usually, \code{n = 8} seems to be a
good choice.}

\item{H}{A list of vertices that are the inner boundary points,
default set to '\code{NULL}' if there is no holes.}
}
\value{
\item{V}{an \code{N} by two matrix that lists vertices with the \code{i}th
row storing in Cartesian coordinates for the
\code{i}th vertex. \code{N} is the number of vertices.}
\item{Tr}{a \code{K} by three matrix that each row represents one triangle.
All the elements are the integers that stand for the indices of vertices in \code{V}.}
}
\description{
This function triangulates the polygonal domain by using Delaunay Triangulation.
}
\details{
In the function, we firstly get grid points inside and on the boundary of
the polygon with extreme points \code{Pt} and interior holes defined by \code{H}. Then delaunay triangulation
is used to generate triangulations by using the grid points.
And lastly we delete triangles within the holes or outside the boundary of the region.
}
\examples{
# rectangular domain
bb=rbind(c(0,0),c(1,0),c(1,1),c(0,1))
VT=TriMesh(bb,2)

# irregular domain
data("horseshoe")
TriMesh(horseshoe,n=8)

# region with holes
data("BMP")
TriMesh(BMP$bound,n=16,list(as.matrix(BMP$H1),as.matrix(BMP$H2)))

data("mymontreal")
plot(mymontreal$bound[,1],mymontreal$bound[,2])
TriMesh(mymontreal$bound,5)
TriMesh(mymontreal$bound,18,list(mymontreal$H1,mymontreal$H2))
}
